Move consumerfanout package to internal (#3207)
