+++ service/internal/fanoutconsumer/consumer.go
